Better support for numeric lists
Lists like [0.0,1.0] are now read correctly, they are no longer read as
int arrays. (int[] cannot be cast to NBTTagList<NBTTagFloat>)

Empty lists like [] are no longer read at all - it is unknown if an int
array or NBTTagList is expected. Luckily NBTTagCompound will provide the
mod with an empy array or list of the correct when it asks for one and
no value is present.