Different saving check
Previously, the plugin only saved chests that have been clicked. This
didn't work nice with other plugins modifying the contents of a chest.

This system has been replaced by a new system, where the plugin looks if
the contents of the chest have been changed. It does this by comparing
hash codes.